+++ apsbss/tests/_core.py
@@ -1,7 +1,6 @@
 """Common support for testing."""
 
 import pathlib
-import socket
 import time
 import uuid
 
@@ -11,11 +10,6 @@ BSS_TEST_IOC_PREFIX = f"tst{uuid.uuid4().hex[:7]}:bss:"
 SRC_PATH = pathlib.Path(__file__).parent.parent
 TEST_DATA_PATH = pathlib.Path(__file__).parent / "data"
 CREDS_FILE = TEST_DATA_PATH / "dev_creds.txt"
-
-
-def is_aps_workstation():
-    """Is this workstation on the APS network?"""
-    return socket.getfqdn().endswith(".aps.anl.gov")
 
 
 def wait_for_IOC(delay=0.05):