+++ CHANGES.rst
@@ -44,6 +44,7 @@ describe the future plans.
    * Add support for direct access to read IS database.
    * Add Server class that chooses between DM or IS interface.
    * Integer timestamp PVs for ESAF start & end and Proposal start, end, & submitted.
+   * User can override default DM URL by setting an environment variable.
 
    Maintenance
    -----------